Dev-191-Adding cookiecutter branch
