--- CHANGELOG.md
@@ -1,15 +1,0 @@
-# Changelog
-
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
-Syntax:
-
-```text
-## [version] - YYYY - MM - DD
-### Added/Removed/Changed/Deprecated/Fixed/Security
-```
-
-## [Unreleased]
