+++ cookiecutter.json
@@ -1,0 +1,10 @@
+  
+{
+    "full_name": "First Last",
+    "email": "user@bank-banque-canada.ca",
+    "github_owner": "aecurrency",
+    "project_name": "Template Python Project",
+    "project_slug": "{{ cookiecutter.project_name.lower().replace(' ', '_').replace('-', '_') }}",
+    "project_short_description": "A Starter Project Template for Python.",
+    "version": "0.1.0"
+  }
