--- mkdocs.yml
@@ -1,26 +1,0 @@
-site_name: Template Python Project Docs
-theme:
-  name: material
-markdown_extensions:
-  - abbr
-  - admonition
-  - attr_list
-  - footnotes
-  - meta
-  - pymdownx.betterem:
-      smart_enable: all
-  - pymdownx.details
-  - pymdownx.emoji:
-      emoji_index: !!python/name:materialx.emoji.twemoji
-      emoji_generator: !!python/name:materialx.emoji.to_svg
-  - pymdownx.highlight
-  - pymdownx.inlinehilite
-  - pymdownx.keys
-  - pymdownx.smartsymbols
-  - pymdownx.snippets
-  - pymdownx.superfences
-  - pymdownx.tabbed
-  
-plugins:
-  - search
-  - mkdocstrings
