--- pyproject.toml
@@ -1,49 +1,0 @@
-[tool.poetry]
-name = "template-python-project"
-version = "0.1.0"
-description = ""
-authors = ["Lucas Melin <mell@bank-banque-canada.ca>"]
-
-[tool.poetry.dependencies]
-python = "^3.6"
-
-[tool.poetry.dev-dependencies]
-coverage = {extras = ["toml"], version = "^5.5"}
-black = "^20.8b1"
-flake8 = "^3.9.0"
-flake8-bandit = "^2.1.2"
-flake8-black = "^0.2.1"
-flake8-bugbear = "^21.4.3"
-flake8-docstrings = "^1.6.0"
-flake8-import-order = "^0.18.1"
-pytest = "^6.2.3"
-pytest-cov = "^2.11.1"
-pytest-mock = "^3.5.1"
-safety = "^1.10.3"
-xdoctest = "^0.15.4"
-rope = "^0.19.0"
-nox = "^2020.12.31"
-mkdocs-material = "^7.1.3"
-mkdocstrings = "^0.15.0"
-flake8-comprehensions = "^3.4.0"
-pep8-naming = "^0.11.1"
-bandit = "^1.7.0"
-
-[build-system]
-requires = ["poetry-core>=1.0.0"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.coverage.paths]
-source = ["src", "*/site-packages"]
-
-[tool.coverage.run]
-branch = true
-source = ["src"]
-
-[tool.coverage.report]
-show_missing = true
-fail_under = 60
-
-[tool.isort]
-profile = "black"
-src_paths = ["src", "tests"]
