--- src/template_python_project/__init__.py
@@ -1,2 +1,0 @@
-"""Template Python Project."""
-__version__ = "0.1.0"
