--- tests/test_init.py
@@ -1,7 +1,0 @@
-"""Test cases for the init module."""
-import template_python_project
-
-
-def test_version():
-    """Verify the package version."""
-    assert template_python_project.__version__ == "0.1.0"
