+++ {{cookiecutter.project_slug}}/.flake8
@@ -1,0 +1,7 @@
+[flake8]
+select = B,B9,BLK,C,C4,D,E,F,S,W
+ignore = E203,W503,E501,B008,D202,D412
+per-file-ignores = tests/*:S101,D100,D103
+max-line-length = 88
+max-complexity = 10
+docstring-convention = google
