+++ {{cookiecutter.project_slug}}/.pre-commit-config.yaml
@@ -1,0 +1,17 @@
+repos:
+  - repo: local
+    hooks:
+      - id: black
+        name: black
+        description: "Black: The uncompromising Python code formatter"
+        entry: poetry run black
+        language: system
+        require_serial: true
+        types: [python]
+      - id: flake8
+        name: flake8
+        description: "`flake8` is a command-line utility for enforcing style consistency across Python projects."
+        entry: poetry run flake8
+        language: system
+        types: [python]
+        require_serial: true
