+++ {{cookiecutter.project_slug}}/docs/changelog.md
@@ -1,0 +1,1 @@
+--8<-- "CHANGELOG.md"
