+++ {{cookiecutter.project_slug}}/src/{{cookiecutter.project_slug}}/__init__.py
@@ -1,0 +1,2 @@
+"""Top-level package for {{ cookiecutter.project_name }}."""
+__version__ = '{{ cookiecutter.version }}'
