+++ {{cookiecutter.project_slug}}/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Template {{ cookiecutter.project_name }} tests."""
