+++ {{cookiecutter.project_slug}}/tests/test_{{cookiecutter.project_slug}}.py
@@ -1,0 +1,7 @@
+"""Test cases for `{{ cookiecutter.project_slug }}` package."""
+import {{cookiecutter.project_slug}}
+
+
+def test_version():
+    """Verify the {{ cookiecutter.project_slug }} package version."""
+    assert {{cookiecutter.project_slug}}.__version__ == "{{ cookiecutter.version }}"
